# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/login": {
            "post": {
                "description": "Login with username and password to get a token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User login info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Register a new user with username, password and role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User registration",
                "parameters": [
                    {
                        "description": "User registration info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "api.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "Password 是用于登录的密码。\n验证规则：必填 (required)。",
                    "type": "string"
                },
                "username": {
                    "description": "Username 是用于登录的用户名。\n验证规则：必填 (required)。",
                    "type": "string"
                }
            }
        },
        "api.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "Token 是一个 JWT (JSON Web Token)，客户端后续需要用它来进行身份认证。",
                    "type": "string"
                }
            }
        },
...
        "api.RegisterRequest": {
            "type": "object",
            "required": [
                "password",
                "role",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "Password 是用户注册时设置的密码。\n验证规则：必填 (required)，最小长度为 6 个字符 (min=6)。",
                    "type": "string",
                    "minLength": 6
                },
                "role": {
                    "description": "Role 是用户注册时指定的角色。\n验证规则：必填 (required)，且值必须是 \"Applicant\" 或 \"Approver\" 两者之一 (oneof=Applicant Approver)。",
                    "type": "string",
                    "enum": [
                        "Applicant",
                        "Approver"
                    ]
                },
                "username": {
                    "description": "Username 是用户注册时期望的用户名。\n验证规则：必填 (required)，最小长度为 4 个字符 (min=4)。",
                    "type": "string",
                    "minLength": 4
                }
            }
        },
...
        "api.UserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID 是用户的唯一标识符 (UUID 格式的字符串)。",
                    "type": "string"
                },
                "role": {
                    "description": "Role 是用户的角色。",
                    "type": "string"
                },
                "username": {
                    "description": "Username 是用户的名称。",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}